+++ src/app/services/api-objects/resource-response-formats/src/incoming-item-json.ts
@@ -1,0 +1,2 @@
+export class IncomingItemJson {
+}
