+++ src/app/services/api-objects/resource-response-formats/src/resdata-json.ts
